Combine CoreData managed classes into single files
