+++ FeedStoreChallenge/Helpers/ManagedCache+CoreDataClass.swift
@@ -11,4 +11,61 @@ import Foundation
 import CoreData
 
 @objc(ManagedCache)
-public class ManagedCache: NSManagedObject {}
+public class ManagedCache: NSManagedObject {
+	@nonobjc public class func fetchRequest() -> NSFetchRequest<ManagedCache> {
+		return NSFetchRequest<ManagedCache>(entityName: "ManagedCache")
+	}
+
+	@NSManaged var timestamp: Date?
+	@NSManaged var feed: NSOrderedSet
+}
+
+extension ManagedCache {
+	var localFeed: [LocalFeedImage] {
+		return feed.compactMap { ($0 as? ManagedFeedImage)?.local }
+	}
+
+	static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
+		let request = NSFetchRequest<ManagedCache>(entityName: entity().name ?? "ManagedCache")
+		request.returnsObjectsAsFaults = false
+		return try context.fetch(request).first
+	}
+
+	static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
+		try find(in: context).map(context.delete)
+		return ManagedCache(context: context)
+	}
+}
+
+// MARK: Generated accessors for feed
+extension ManagedCache {
+	@objc(insertObject:inFeedAtIndex:)
+	@NSManaged public func insertIntoFeed(_ value: ManagedFeedImage, at idx: Int)
+
+	@objc(removeObjectFromFeedAtIndex:)
+	@NSManaged public func removeFromFeed(at idx: Int)
+
+	@objc(insertFeed:atIndexes:)
+	@NSManaged public func insertIntoFeed(_ values: [ManagedFeedImage], at indexes: NSIndexSet)
+
+	@objc(removeFeedAtIndexes:)
+	@NSManaged public func removeFromFeed(at indexes: NSIndexSet)
+
+	@objc(replaceObjectInFeedAtIndex:withObject:)
+	@NSManaged public func replaceFeed(at idx: Int, with value: ManagedFeedImage)
+
+	@objc(replaceFeedAtIndexes:withFeed:)
+	@NSManaged public func replaceFeed(at indexes: NSIndexSet, with values: [ManagedFeedImage])
+
+	@objc(addFeedObject:)
+	@NSManaged public func addToFeed(_ value: ManagedFeedImage)
+
+	@objc(removeFeedObject:)
+	@NSManaged public func removeFromFeed(_ value: ManagedFeedImage)
+
+	@objc(addFeed:)
+	@NSManaged public func addToFeed(_ values: NSOrderedSet)
+
+	@objc(removeFeed:)
+	@NSManaged public func removeFromFeed(_ values: NSOrderedSet)
+}
